+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## 2.9.0 (Unreleased)
-
+* ⚠️ Updated to Kotlin 2.1.10 ⚠️
+* ⚠️ Updated project to support Kotlin Multiplatform ⚠️
+    * Only the Android target is currently supported, but the project is now set up to support other targets in the future
+* Fixed a possible memory leak bug with `ComposableDestinationOwner`
 
 ## 2.8.3
 * Resolved a bug with animation changes to `BottomSheetDestination` that caused animation snapping for these destinations
@@ -10,7 +13,6 @@
 * Removed deprecated DialogDestination and BottomSheetDestination interfaces, and associated functions. Please use the Composable `DialogDestination` and `BottomSheetDestination` functions instead. Example usage can be found in the test application.
 * Deprecated the `OverrideNavigationAnimations` function that does not take a content lambda, in favour of the version that does take a content lambda.
 * `ModalBottomSheetState.bindToNavigationHandle` no longer overrides navigation animations.
-* Updated animation internals in preparation for predictive back navigation
 
 ## 2.8.1
 * Fixed a bug with ComposableDestinationSavedStateOwner that was causing lists of primitives (such as List<Int>) to not get saved/restored correctly